Split CVXQUAD cones into toqito/cones/ subpackage
The CVXQUAD-derived matrix-cone and entropy modules (geometric_mean*,
trace_matrix_power, lieb_ando, operator_relative_entropy_epi_cone,
trace_matrix_log, ln_quantum_entropy) lived in matrix_ops but
*require* matrix_props for input validation, which forced
matrix_ops to depend on matrix_props. Several matrix_props modules
(is_orthonormal, is_block_positive via sk_norm, is_ldoi)
transitively depend on state_props or channels, which in turn
import back through matrices.pauli -> matrix_ops, closing a
load-time cycle.

Recent PRs (#1559, #1560) deferred imports with '# noqa: PLC0415'
to mask the cycle one back-edge at a time. Each new CVXQUAD module
re-surfaced it elsewhere. Move the offending modules out of
matrix_ops so matrix_ops is truly low-level (no matrix_props
dependency), and the cycle disappears structurally.

- New subpackage toqito/cones/ holds the 8 CVXQUAD modules + the
  shared _utils (formerly _cone_utils) helper.
- matrix_ops/__init__.py no longer eagerly imports them; users now
  do 'from toqito.cones import ...'.
- state_props.quantum_relative_entropy retargets its imports to
  toqito.cones.
- matrix_props/__init__.py, is_block_positive, is_ldoi, and
  test_is_block_positive lose their cycle-related band-aids (deferred
  imports, lazy __getattr__, monkeypatch indirection). Their
  matrix_props imports go back to plain top-level form.
- is_orthonormal keeps the dead-code cleanup from #1559 (the
  redundant is_mutually_orthogonal call was already unnecessary
  before the cycle work).

Verified locally: docs-deploy entry path (toqito.matrices import
chain) imports cleanly; matrix_props + state_props + cones +
channel_metrics + matrix_ops tests all pass.
