+++ toqito/matrix_ops/__init__.py
@@ -14,13 +14,3 @@ from toqito.matrix_ops.tensor_unravel import tensor_unravel
 from toqito.matrix_ops.partial_trace import partial_trace
 from toqito.matrix_ops.partial_transpose import partial_transpose
 from toqito.matrix_ops.null_space import null_space
-from toqito.matrix_ops.geometric_mean import geometric_mean
-from toqito.matrix_ops.geometric_mean_epi_cone import geometric_mean_epi_cone
-from toqito.matrix_ops.geometric_mean_hypo_cone import geometric_mean_hypo_cone
-from toqito.matrix_ops.trace_matrix_power import trace_matrix_power
-from toqito.matrix_ops.lieb_ando import lieb_ando
-from toqito.matrix_ops.operator_relative_entropy_epi_cone import (
-    operator_relative_entropy_epi_cone,
-)
-from toqito.matrix_ops.trace_matrix_log import trace_matrix_log
-from toqito.matrix_ops.ln_quantum_entropy import ln_quantum_entropy
