+++ toqito/state_props/quantum_relative_entropy.py
@@ -7,12 +7,12 @@ import cvxpy
 import numpy as np
 from scipy.linalg import logm
 
-from toqito.matrix_ops._cone_utils import _require_square_2d
-from toqito.matrix_ops.ln_quantum_entropy import ln_quantum_entropy
-from toqito.matrix_ops.operator_relative_entropy_epi_cone import (
+from toqito.cones._utils import _require_square_2d
+from toqito.cones.ln_quantum_entropy import ln_quantum_entropy
+from toqito.cones.operator_relative_entropy_epi_cone import (
     operator_relative_entropy_epi_cone,
 )
-from toqito.matrix_ops.trace_matrix_log import trace_matrix_log
+from toqito.cones.trace_matrix_log import trace_matrix_log
 from toqito.matrix_props import is_hermitian, is_positive_semidefinite
